Разделение проекта на модули
