+++ js/ad-form.js
@@ -1,0 +1,1 @@
+//Пустой модуль на будущее
