+++ js/utils/get-random-positive-float.js
@@ -20,3 +20,4 @@ function getRandomPositiveFloat (a, b, digits = 1) {
   // указать требуемое количество знаков после точки
   return result.toFixed(digits);
 }
+getRandomPositiveFloat (0,1);
